pyproject.toml
[project]
name = "fitnessllm-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Santosh Gummidipundi"}
]
requires-python = "3.12.2"
dependencies = [
    "langchain-deepseek-official (>=0.1.0.post1,<0.2.0)",
    "notebook (>=7.3.2,<8.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "jupyter-server (==2.12.5)",
    "langchain (>=0.3.18,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langchain-core (>=0.3.35,<0.4.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "sqlalchemy-bigquery (>=1.12.1,<2.0.0)",
    "google-cloud-bigquery-storage (>=2.28.0,<3.0.0)",
    "google-cloud-bigquery (>=3.29.0,<4.0.0)",
    "torch (>=2.6.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
